Possible improvement in WindowsPhone for multiple images
For each image, the code checks if the file (placeholder, error, etc) exists, through StorageFile.GetFileFromPathAsync statement.
My proposal is to create a local dictionary where the source is the key, where to store the result of the check just one time per source.
With this approach I've optimized the loading time of a list view with 25 images by reducing the load time to more than 400%