Introduce fallback resolver
MAUI generally uses some different methods for MauiImage type files and MauiAssets.  It's not always clear now when you give a filename such as "sample.svg" if it's going to be found by a specific data resolver.

For instance, on Android, it could be a Resource (so, ResourceDataResolver), or it could be an Asset (so BundleDataResolver).  On the other hand, a Filepath source type could be any one of those data resolvers in addition to a FileDataResolver.

This wraps a set of resolvers and tries them in order and falls back through them all if any fail.